port freeRuns to their own file, make withRun wrapper
